chore: better code
